Add files via upload
First, I created an S3 bucket on AWS to store the Terraform state file. Then, I added a provider section to install the plugin. After that, I created a virtual private subnet "my_vpc". Inside this VPC, I provisioned two subnets. One is a public subnet "pub_sub" containing an EC2 instance "APP_instance". The second subnet is a private subnet "priv_sub" which contains an EC2 instance "DB_instance". I provisioned an internet gateway "my_gw" to allow the instance in the VPC to communicate with the internet and then associated it with the route table.